requirements.txt
MetaTrader5==5.0.45
pandas==2.2.0
numpy==1.26.0
schedule==1.2.1
python-dotenv==1.0.0
